initial placeholder commit, for module installation
The previous commit had files in the top level of the repo in a
flat hierarchy, which isn't optimal for module structure. We
create a new pyani_plus subdirectory to hold the source code, and
add two subdirectories: snakemake, for snakemake wrappers, and
workflows, for the snakemake workflow files This structure/
naming scheme might change later.

We add __init__.py files to each importable module/submodule, and
update requirements.txt to include snakemake. We update the
pyproject.toml file to point to the appropriate source base, and
update the ruff configuration.