Helper module of get-version functions
Based on a mixture of the existing get_version functions
in pyani which provides the regex based version parsing
(e.g. pyani/anim.py and pyani/fastani.py) and my similar
code in THAPBI PICT:

https://github.com/peterjc/thapbi-pict/blob/v1.0.13/thapbi_pict/versions.py

Then tweaked to pass the strict ruff checks and refactored
to reduce duplicated code.

Unlike pyani, this combines stdout+stderr when capture.
This doesn't use the legacy subprocess getoutput(...) or
getstatusoutput() functions which are documented as less
secure (e.g. they run with shell=True).

Unlike pyani, this delegates all error handling to the
"which" call which we do anyway to find out the full path.

The functions return a named tuple of the executable path
and the version string.