Use workflow param to pass explicit tool binary paths
This lets us determine the paths to the assorted
binary tools like nucmer before callign snakemake,
and record their location and version etc.

Appears the config dict must define all the keys used
in a workflow file, not just those in any single rule.
In our case most of the values are the same, but for
the output directory which is test dependent.