Record arch and os in the Configuration table
This is the value from 'arch' or 'uname -m' for the CPU architecture
or machine name (e.g. 'arm64') and 'uname -s' for the operating
system (e.g. 'Darwin' for the macOS Kernal).

The rational is the potential for cross-platform differences may
be enough for a different ANI value (e.g. floating point differences
in C libraries or in CPU families).

Note the test is getting a bit long, hitting PLR0915 from ruff.