Cache identities, query-cov, etc in DB as JSON
Using pandas JSON with orient="split" for size.
Could improve on this further as it wastefully
records the hashes twices as both index (rows)
and column names...

With the exception of the Hadamard matrix, the caching
method does not do any calculations - just collecting
N^2 matrix elements from the comparisons table.