Free large arrays as early as possible.
With N=1000 and a million comparisons, simple benchmarking
puts the caching at about 30s but nearly 2GB of RAM.

With N=1500, about 90s and 4GB of RAM in Activity Monitor.

In principle we could reduce the peak RAM by doing the matrices
in batches. Worst case is Hadamard matrix as needs both the
identity and coverage present. Doing that in one pass, and the
rest in a a second pass does work - but it takes longer and
doesn't seem to help reduce the RAM usage.