Commit in the add helper function; avoid locks
These changes were driven by testing the fastANI snakemake
workflow which was calling the private CLI from each worker
to log the comparisons.

This was also recording the genomes, which tended to clash.
We may want to pre-record all the genomes before starting
the comparisons which would avoid this source of conflict.
I do not anticipate a similar issue for the comparisons.