OP073605.1 as third virus for test case
Leighton wanted 3 or 4 viruses, not just two.

Angelika identified this (OP073605.1) and a fourth
example (OP075722.1 MAG: Bacteriophage sp. isolate
3686_45362, partial genome) as aligning with the
first two - see #95.

This was downloaded from the NCBI, and given the
extension .fasta to rather than .fna to ensure the
test set covers a range of extensions (see the
bacterial example).