refractor dnadiff workflow, generate new fixtures, update tests
Currently, the dnadiff workflow generates .rdiff files that are be used
for calculations related to the reference. A while ago, we had a
conversation about reporting values for the query instead. In this
commit, I:

- Made changes to the dnadiff workflow to generate qdiff files rather than
rdiff files.
- Modified generate_fixtures to generate qdiff files.
- Provided .qdiff files.
- Changed the arguments for show-coords so that they match the arguments
implemented in dnadiff. Previous arguments would lead to the same result,
but new arguments will help with parsing mcoord files.