Make DB logging API stricter; rename add_comparison etc
Renames add_comparison -> db_comparison etc; adds
explicit create boolean flag for adding configuration
or genome.

This changes the API to make adding configuration and
genome rows configurable.

Currently adding a comparison would implicitly add the
configuration and genome rows if not already there -
which may be undesirable (e.g. mismatch between the
intended configuration already in the DB, and what
snakemake is trying to use instead which if wrong will
not already exist).

With these changes log_comparison and the log_<method>
functions like log_fastani will now REQUIRE that the
database exists and contains the configuration and
genomes already.

This ensures our snakemake workflows WILL FAIL if they
are trying to write to a non-existent configuration due
to an error in propagating the configuration settings.