Update progress-bar from child thread; snakemake on main
Previously we updated the progress bar on the main thread,
with the single 'black-box' function call to snakemake on
a child thread. That worked, but when snakemake failed we
didn't get to see the traceback.

This way round we do see the snakemake traceback (e.g. by
pressing ctrl+c to abort it).