Call snakemake via snakemake.cli top level functions only.
This essentially means we control snakemake almost as a we
would at the command line, which means in particular we now
automatically have snakemake's built in profile support.

Looking at the source code of snakemake v8.20.3, they way
that profiles (YAML configuration files for things like local
HPC specific settings) are handled is as part of parsing the
command line. This means we'd have to mimic some of that logic
if we wanted to pull SLURM settings from the YAML and apply it
to how we called the snakemake API. That looked tricky.

This approach of calling snakemake (almost) via their CLI
means we can drop any details about the more internal API.