Use rich.table for the list-runs output
This includes a "Comparisons" column showing
done/total=n*n, where "done" is the number of
comparisons recorded for that run, total is the
number expected, equal to n-squared for n genomes.
This is colour coded, red if done is zero, yellow
when done != total, and green for done=total.

We may want to offer a choice of output format.
Legacy pyani offered tab separated, Excel, html.
Might be useful to offer Markdown table output?