Add fasta_filename to the runs-genomes linker table
This is being done as a required field, which means
the old helper methods for adding runs to a genome,
or genomes to a run can't work. They are now read-only.

Note we're storing the filename alone without its path
information - that is already in the run table entry
(and by design is shared for all FASTA in a run).

This is motivated by the planned resume functionality.