When log pairwise ANI, pass configuration via run_id
Giving the run_id indirectly gives the DB configuration, and as a
single argument is easier/simpler when writing the workflows
compared to remembering to set all the (possibly optional)
available parameters.

This also lets us avoid recomputing the FASTA file MD5 checksums
as they are available from the run table as well.