sourmash resume support
Because we compute all-vs-all in a single call to sourmash compare
(for speed in the typical clean slate situation), if we are to
resume a partial run - or indeed work with any pre-existing
values (e.g. user previously had N genomes and all NxN are in the DB,
but in a new run expands to N+1 genomes), must wastefully recompute
them but not try to re-add them to the database.