Spinner for sourmash etc rather than progress bar.
The progress bar wouldn't actually update until the end as
this was an all-in-one computation and bulk import, so
visually not ideal.

Worse, the useless progress bar was checking the database
and I did in testing with sourmash brakewater see the
logging fail with a conflict - possibly due to this.