Refactor, introducing private CLI compute_column
My vision here is (nearly?) all the methods can be computed by column,
and this allows them to share the basic setup like working out which
queries are needed for the given subject.

Currenly only using this for fastANI and ANIb, both of which as setup
have no prior task dependencies (ANIb now computes the database for
the subject here, and fragments each query on demand). This means they
could now in principle share the same snakemake workflow definition, a
generic computer-by-column workflow (which could also be done as a
cluster array job).