Avoid race condition in how ANIb fragmented FASTA are made
If running ANIb with an explicitly configured temp-dir, all the
works for each subject (column) will write to the same place.
Previously they would be sharing the same fragmeneted FASTA files
leading to read/write races.

It might be better to avoid a temp file at all, and write to
the blastn stdin?