Test wrapped subprocess.check_output error message
Moved from using sys.exit(returncode) to sys.exit(msg) in order to
be able to test the error message (and had to accept always code 1).

Note preceeding commit, pytest capsys was not working for me to test
the error message by looking in stderr instead.

Sample command for testing:

❯ python -c "from pyani_plus.utils import check_output; check_output(['blastn', '-task','undefined'])"; echo "Return code $?"