Start testing query coverage with artifical example.
Using a simple example of a small contig, a large contig, and both contigs.
The small and large contig have nothing in common, so barring noise in the
self-vs-self we can expect identity values of 99~100% or null (between the
small and large), while for query coverage we expect 100%, 30% (small vs
both), 70% (large vs both), or null (between small and large).