Start testing query coverage with artifical example.
Using a simple example of a small contig, a large contig, and both contigs.
The small and large contig have nothing in common, so barring noise in the
self-vs-self we can expect identity values of 99~100% or null (between the
small and large), while for query coverage we expect 100%, 30% (small vs
both), 70% (large vs both), or null (between small and large).

This highlights that ANIb is still recording zero rather than null for
failed comparisons, and fastANI is actually failing to complete on two
of the comparisons.

This "works" for sourmash as well (confirming the coverage matrix is
correctly orientated vs the transpose), although here the k-mer based
approach doesn't give use 30% and 70%, and we've had to use an extreme
scalling to get any match at all.