Record sourmash query-containment as query-coverage
Will now always record max-containment as the ANI (which was the
default mode), and record containment via the query-coverage matrix
along side it.

To do this, we no longer run sourmash in --max-containment mode,
but just in --containment mode and infer the max value for each
query/subject pair.

This drops --mode as a user-facing option.

Note: May need to transpose the sourmash.csv before loading it
as our query-coverage values.