Reduce blastn TSV output to just the 6 fields parsed + subject name
This cuts the size of the intermediate TSV output by about half,
so will hopefully have a small improvement on the runtime.

Note: Did not check-in the revised NJS files. The datastamp
change is not of interest, but curiously the "bytes-total"
values on macOS are much larger than on Linux. Perhaps this
is linked to the file system?