Log ANIm, dnadiff, and ANIb in batches
In order to avoid locking errors writing to SQLite3 on a network drive
we had switched to recording the comparisons for these methods a whole
column at a time - and this has been tested with up to 400 workers on
400 genomes.

However, this change means the main application shows zero progress
until the first column completes - which is not a good user experience.
We therefore want to log DB output occasionally, in order to have the
main thread progress bar update, but not too often and not all at once.

Assuming all the pairwise comparisons take the same time, this design
would mean at each round a different column worker would log its progress.
This would mean regular updates until all the columns complete, although
at any one moment in addition to what has been logged and shown on the
progress bar, there will be lots of completed calculations cached and
waiting to be logged.