Catch SIGINT / KeyboardInterrupt in ANIm/dnadiff/ANIb & log to DB
These three methods are the slowest and also the only ones
where we are calling the external tools for each pairwise
comparison. This means if the task is interupted (e.g. user
pressing ctrl+c, or SLURM sending SIGINT) we may have some
computed comparisons not yet recorded to the database.

With this change we will attempt to log them to the database.
This may of course lead to locking clashes if multiple
aborted workers are all doing this at the same time, but
we should still record much of the progress which would
otherwise simply be lost.