Make numpy & pandas imports lazy (not wanted for private CLI)
Prompted by Tests/test_interupt.py sometimes failing locally as the
private CLI was still importing pandas when the SIGINT was received.

The only complication here is the use of pd.DataFrame in the type
annotation, which is now handled via an if TYPE_CHECKING condition
and putting the DataFrame annotation in strings ("forward references").