Simple long-form table from run-export
For partial runs will export the available comparisons in long-form,
and then error. i.e. Still no support for partial matrix output.

For long-form the Hadamard is computed on the fly (for the matrix
this takes the cached data-frame from the DB instead).

Currently the long-form query and subject are always labelled with
their MD5 (does not yet follow the --label option and use filename
or stem).