Refactor import external-alignment to run in parallel by column
This makes it look just like the others computed by column (leaving
just sourmash/branchwater as an odd one).

This method is by definition symmetric, so of the N by N comparisons only
half are needed (plus the diagonals which should be 100% identity but we
still need to compute the coverage).

The first column should be a full run (all N comparisons), the second column
can skip one entry leaving N-1 comparisons, etc, down to the final column
being a trivial single self-vs-self comparison.

It would be logical to try splitting the jobs by pairs of columns, first
and last together (first alone if N is odd).