Refactor sourmash compute column into compute tile; call once
This special cases sourmash to be called via compute column with
the special column value of n meaning all columns (could use zero
instead leaving 1 to n for doing a single column). Doing a single
call to sourmash is sensible even for 100s of genomes, attempting
this by column was orders of magnitude slower due to overheads.