Log compute-column interuptions
Not using logger.exception (suggested by ruff TRY004) as the interuption
exception details are not important.

Note this required explicitly setting up the terminal logger for the private CLI,
and this is not using rich with nice colours etc. Typically we will be calling
the private CLI compute-column in quiet mode, so the terminal logging is only
seen for errors anyway - and importing rich is best avoided for the compute jobs.