Call fastANI with batches of queries (not entire column)
This means if the jobs are interrupted we'll have recorded any completed
batches - rather than all the computation thus far being wasted.

It also means more smaller DB writes, which may help with the high level of
DB locking failures seen on Archie West runing a ~1000 genome example. This
means the first batches have multiple logging attempts, and the final logging
after all the computation should be for much less than a full column - and
one hopes will be more likely to succeeed with the three-strikes-and-out.