Start column-specific log for compute-column as early as possible.
This is essential to capture failures to connect to the DB (at which point
if using a subject name we wouldn't actually know the column number - and
so wouldn't be using a per-column log file yet).

i.e. Logical next step is to change the private CLI to only take column numbers.

Also, connect_to_db should take a logger argument.