resolve() != absolute(), and ought to pass absolute cache directory too.
Calling resolve will cleanup any .. in the path (nice) but also dereference
any symlinks (bad if for example the DB path we've been given is actually
a symlink to some fast storage). Since we interpret input FASTA file locations
relative to the DB location as given, should use .absolute() not .resolve() here.

Also, in local testing managed to get sourmash to fail with cache '.' not
being considered in the context of a temp working directory.